Update Lean versioning to diff files (#1)
When new features and bug fixes are added to the REPL, they are not backported to previous Lean versions using the official tag-based approach. Here we choose to decouple Lean versions from the REPL versions by adding a `versions` folder containing diff files to apply when switching to a different Lean version.
The CI is updated accordingly.